+++ src/AVPRCI/CLIArgs.fs
@@ -5,12 +5,14 @@ open Argu
 type PublishArgs =
     | [<AltCommandLine("-d")>] Dry_Run
     | [<ExactlyOnce; AltCommandLine("-k")>] API_Key of string
+    | [<ExactlyOnce; AltCommandLine("-u")>] Base_Url of string
 
     interface IArgParserTemplate with
         member s.Usage =
             match s with
             | Dry_Run          -> """Optional | Default = false | Dry run mode enabled. This will only print a preview of the changes that would be pushed to the package database."""
             | API_Key _        -> """Required | API key for the package database."""
+            | Base_Url _       -> """Required | Base URL of the target package registry, for example https://avpr.nfdi4plants.org."""
 
 
 [<HelpFlags([|"--help"; "-h"|])>]